backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-cloud-firestore==2.23.0
google-cloud-pubsub==2.34.0
google-cloud-bigquery==3.25.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.1
email-validator==2.2.0
shapely==2.0.4
websockets==12.0

# --- Dev / test only ---
pytest==8.3.0
httpx==0.27.0
